Create index-year.php
store a cookie for a period of 1 year. To do this, you can set the cookie expiration date using PHP's setcookie function.

In this example, the cookie named "cookie_name" will be stored with the value "my_cookie" and will expire 1 year from the time it is set. The cookie can only be accessed from the same domain and can only be accessed over HTTPS. Furthermore, the cookie is only accessible by the server, it cannot be accessed by JavaScript in the browser.