Create index.php
If the title is not found in the og:title tag, an alternative may be to extract the page title from the HTML page itself. I will adjust the PHP code to extract the page title if og:title is not found. Here is how you can do this!

In this tweaked PHP code, if the title is not found in the og:title tag, we try to extract the title from the page itself using the <title> tag. If neither og:title nor <title> are found, we return a generic site name. This ensures that there will always be some sort of title displayed, even if it's not page-specific.